add redirect to workouts on connection and reverse on disconnect
