+++ src/components/Layout/index.jsx
@@ -1,8 +1,20 @@
+import { useEffect } from 'react'
 import './styles.css'
 
-import { Outlet } from 'react-router-dom'
+import { Outlet, useNavigate } from 'react-router-dom'
 
-export default function Layout () {
+export default function Layout ({connected}) {
+    const navigate = useNavigate()
+
+    useEffect(()=>{
+        if (!connected){
+            navigate('/')
+        }
+        if (connected) {
+            navigate("/workout")
+          }
+    },[connected])
+
     return (
         <main>
             <Outlet />
