+++ src/components/workouts/RPE/index.jsx
@@ -15,12 +15,12 @@ export default function RPEWorkout({
   setMeasuring,
 }) {
   const [RPE, setRPE] = useState(7);
-  const [pullTime, setPullTime] = useState(3);
-  const [restTime, setRestTime] = useState(180);
+  const [pullTime, setPullTime] = useState(7);
+  const [restTime, setRestTime] = useState(120);
 
   const [resting, setResting] = useState(false);
   // place holder weight while I build out db
-  const [maxWeight, setMaxWeight] = useState(1);
+  const [maxWeight, setMaxWeight] = useState(70);
 
   const { time, setTime, isRunning, start, stop } = useTimer(pullTime);
